# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Pharmacy POS API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /sales:
    post:
      operationId: createSale
      tags: [sales]
      summary: Create sale
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SaleInput"
      responses:
        "201":
          description: Sale created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SaleResponse"
    get:
      operationId: listSales
      tags: [sales]
      summary: List sales
      parameters:
        - name: from
          in: query
          schema:
            type: string
        - name: to
          in: query
          schema:
            type: string
        - name: branch_id
          in: query
          schema:
            type: string
        - name: patient_id
          in: query
          schema:
            type: string
        - name: sale_type
          in: query
          schema:
            type: string
        - name: payment_method
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: per_page
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Sales
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SaleListResponse"
  /sales/kpis:
    get:
      operationId: getSalesKpis
      tags: [sales]
      summary: Get sales KPIs
      parameters:
        - name: branch_id
          in: query
          schema:
            type: string
        - name: date
          in: query
          schema:
            type: string
      responses:
        "200":
          description: KPIs
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SalesKpis"
  /sales/daily-summary:
    get:
      operationId: getDailySummary
      tags: [sales]
      summary: Get daily summary
      parameters:
        - name: branch_id
          in: query
          schema:
            type: string
        - name: date
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Daily summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DailySummary"
  /sales/held:
    get:
      operationId: listHeldSales
      tags: [sales]
      summary: List held sales
      parameters:
        - name: branch_id
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Held sales
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Sale"
  /sales/calculate:
    post:
      operationId: calculateSale
      tags: [sales]
      summary: Calculate sale totals (dry-run)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SaleInput"
      responses:
        "200":
          description: Calculation result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SaleCalculation"
  /sales/{sale_id}:
    get:
      operationId: getSale
      tags: [sales]
      summary: Get sale
      parameters:
        - name: sale_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Sale
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Sale"
  /sales/{sale_id}/refund:
    post:
      operationId: refundSale
      tags: [sales]
      summary: Process refund
      parameters:
        - name: sale_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefundInput"
      responses:
        "200":
          description: Refund processed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Sale"
  /sales/{sale_id}/hold:
    post:
      operationId: holdSale
      tags: [sales]
      summary: Hold sale
      parameters:
        - name: sale_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Held
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Sale"

  # ── SUPPLIERS ──────────────────────────────────────────────────
...
components:
  schemas:
...
    Pagination:
      type: object
      properties:
        page:
          type: integer
        per_page:
          type: integer
        total:
          type: integer
        pages:
          type: integer
        has_next:
          type: boolean
        has_prev:
          type: boolean
      required:
        - page
        - per_page
        - total
        - pages
        - has_next
        - has_prev

    # ── AUTH ──────────────────────────────────────────────────
...
    SaleItem:
      type: object
      properties:
        id:
          type: string
        medicine_id:
          type: string
        medicine_name:
          type: string
        batch_id:
          type: ["string", "null"]
        quantity:
          type: integer
        unit_price:
          type: number
        discount_pct:
          type: number
        discount_flat:
          type: number
        line_total:
          type: number
      required:
        - id
        - medicine_id
        - medicine_name
        - quantity
        - unit_price
        - line_total
    SaleItemInput:
      type: object
      properties:
        medicine_id:
          type: string
        batch_id:
          type: ["string", "null"]
        quantity:
          type: integer
        unit_price:
          type: number
        discount_pct:
          type: number
        discount_flat:
          type: number
      required:
        - medicine_id
        - quantity
        - unit_price
    Sale:
      type: object
      properties:
        id:
          type: string
        invoice_number:
          type: string
        branch_id:
          type: string
        patient_id:
          type: ["string", "null"]
        patient_name:
          type: ["string", "null"]
        prescription_id:
          type: ["string", "null"]
        sale_type:
          type: string
        cashier_id:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/SaleItem"
        subtotal:
          type: number
        discount:
          type: number
        tax:
          type: number
        total:
          type: number
        payment_method:
          type: string
        amount_tendered:
          type: ["number", "null"]
        change:
          type: ["number", "null"]
        status:
          type: string
        created_at:
          type: string
      required:
        - id
        - invoice_number
        - branch_id
        - sale_type
        - cashier_id
        - items
        - subtotal
        - discount
        - tax
        - total
        - payment_method
        - status
        - created_at
    SaleInput:
      type: object
      properties:
        branch_id:
          type: string
        patient_id:
          type: ["string", "null"]
        prescription_id:
          type: ["string", "null"]
        sale_type:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/SaleItemInput"
        discount_on_invoice:
          type: number
        discount_pct_on_invoice:
          type: number
        tax_rate:
          type: number
        payment_method:
          type: string
        amount_tendered:
          type: ["number", "null"]
        notes:
          type: ["string", "null"]
      required:
        - branch_id
        - sale_type
        - items
        - payment_method
    SaleResponse:
      type: object
      properties:
        sale:
          $ref: "#/components/schemas/Sale"
        warnings:
          type: array
          items:
            type: object
            properties:
              type:
                type: string
              severity:
                type: string
              medicines:
                type: array
                items:
                  type: string
            required:
              - type
              - severity
              - medicines
        stock_updated:
          type: array
          items:
            type: object
            properties:
              medicine_id:
                type: string
              new_qty:
                type: integer
            required:
              - medicine_id
              - new_qty
      required:
        - sale
        - warnings
        - stock_updated
    SaleListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Sale"
        pagination:
          $ref: "#/components/schemas/Pagination"
      required:
        - data
        - pagination
    SalesKpis:
      type: object
      properties:
        branch_id:
          type: ["string", "null"]
        date:
          type: string
        today_revenue:
          type: number
        today_transactions:
          type: integer
        avg_basket:
          type: number
        top_medicine:
          type: ["string", "null"]
        top_cashier:
          type: ["string", "null"]
        refunds_today:
          type: number
        net_revenue:
          type: number
      required:
        - date
        - today_revenue
        - today_transactions
        - avg_basket
        - refunds_today
        - net_revenue
    DailySummary:
      type: object
      properties:
        date:
          type: string
        branch_id:
          type: ["string", "null"]
        total_sales:
          type: number
        total_transactions:
          type: integer
        cash_sales:
          type: number
        card_sales:
          type: number
        credit_sales:
          type: number
        total_refunds:
          type: number
        net_revenue:
          type: number
        rx_sales:
          type: integer
        otc_sales:
          type: integer
      required:
        - date
        - total_sales
        - total_transactions
        - cash_sales
        - card_sales
        - credit_sales
        - total_refunds
        - net_revenue
    SaleCalculation:
      type: object
      properties:
        subtotal:
          type: number
        discount:
          type: number
        tax:
          type: number
        total:
          type: number
        change:
          type: ["number", "null"]
        warnings:
          type: array
          items:
            type: string
      required:
        - subtotal
        - discount
        - tax
        - total
        - warnings
    RefundInput:
      type: object
      properties:
        items:
          type: array
          items:
            type: object
            properties:
              sale_item_id:
                type: string
              quantity:
                type: integer
              reason:
                type: string
            required:
              - sale_item_id
              - quantity
              - reason
        refund_method:
          type: string
        restock:
          type: boolean
        notes:
          type: ["string", "null"]
      required:
        - items
        - refund_method
        - restock

    # ── SUPPLIERS ──────────────────────────────────────────────────
...